remove Carthage and replace with SPM

+++ MyBusMapSwiftUI/Views/PlacesSearch.swift
@@ -17,7 +17,7 @@ struct PlacesSearch: UIViewControllerRepresentable {
     func makeUIViewController(context: Context) -> GMSAutocompleteViewController {
         let autocompleteController = GMSAutocompleteViewController()
         // Specify the place data types to return.
-        let fields: GMSPlaceField = GMSPlaceField(rawValue: UInt(GMSPlaceField.name.rawValue) | UInt(GMSPlaceField.coordinate.rawValue))
+        let fields: GMSPlaceField = GMSPlaceField(rawValue: UInt64(UInt(GMSPlaceField.name.rawValue) | UInt(GMSPlaceField.coordinate.rawValue)))
         autocompleteController.placeFields = fields
         
         // Specify a filter.
